Get more explicit about requirements to build launchwin
I worked a bunch on understanding how the launcher works and gets built.
In doing so I discovered that I can't build it with python 3.5 because
we're using py2exe which doesn't support python 3.5. I also discovered
there's some requirements that aren't explicitly installed and there's not
a good way to install them on behalf of the user. So, at the least we can
let them know about the problems and make it easier for them to get the
right output and decent errors

This also changes the launch builder to take in a parameter that is the
location of the python3.4 executable. It defaults to the default location
of the install on windows

+++ launchers/make_launcher.ps1
@@ -1,10 +1,26 @@
-param([string]$arch = $ENV:PROCESSOR_ARCHITECTURE)
+param([string]$python = "C:\Python34\python.exe", [string]$arch = $ENV:PROCESSOR_ARCHITECTURE)
 Write-Host "Starting build of windows launcher on $arch"
-Remove-Item launchwin -Recurse
+If (Test-Path launchwin) {
+    Remove-Item launchwin -Recurse
+}
+If (Test-Path launch.exe) {
+    Remove-Item launch.exe
+}
+If (Test-Path launchwin.zip) {
+    Remove-Item launchwin.zip
+}
 Write-Host "Building launcher with py2exe"
-& python.exe setup.py py2exe
+& $python setup.py py2exe
+If (!(Test-Path launchwin\launch.exe)) {
+    Write-Host "Unable to create launch\launch.exe. Exiting"
+    exit 1
+}
 Write-Host "Zipping up launcher"
-& python.exe dist2zip.py
+& $python dist2zip.py
+If (!(Test-Path launchwin.zip)) {
+    Write-Host "Unable to create launchwin.zip. Exiting"
+    exit 1
+}
 switch ($ENV:PROCESSOR_ARCHITECTURE)
 {
 	"AMD64" {$OUTPUT = "launchwin-amd64.zip"}
